+++ RandomSkunk.Results.UnitTests/Linq_extension_methods.cs
@@ -64,10 +64,11 @@ public class Linq_extension_methods
         [Fact]
         public void Given_all_success_results_Returns_success_result()
         {
-            var result = from n in 1.ToResult()
-                         let n2 = n * 2
-                         from s in $"n2: {n2}".ToResult()
-                         select s;
+            var result =
+                from n in 1.ToResult()
+                let n2 = n * 2
+                from s in $"n2: {n2}".ToResult()
+                select s;
 
             result.IsSuccess.Should().BeTrue();
             result.Value.Should().Be("n2: 2");
@@ -76,10 +77,11 @@ public class Linq_extension_methods
         [Fact]
         public void Given_early_fail_result_Returns_fail_result()
         {
-            var result = from n in Result<int>.Fail("A")
-                         let n2 = n * 2
-                         from s in $"n2: {n2}".ToResult()
-                         select s;
+            var result =
+                from n in Result<int>.Fail("A")
+                let n2 = n * 2
+                from s in $"n2: {n2}".ToResult()
+                select s;
 
             result.IsFail.Should().BeTrue();
             result.Error.Message.Should().Be("A");
@@ -88,10 +90,11 @@ public class Linq_extension_methods
         [Fact]
         public void Given_late_fail_result_Returns_fail_result()
         {
-            var result = from n in 1.ToResult()
-                         let n2 = n * 2
-                         from s in Result<string>.Fail("B")
-                         select s;
+            var result =
+                from n in 1.ToResult()
+                let n2 = n * 2
+                from s in Result<string>.Fail("B")
+                select s;
 
             result.IsFail.Should().BeTrue();
             result.Error.Message.Should().Be("B");
@@ -100,10 +103,11 @@ public class Linq_extension_methods
         [Fact]
         public void Given_early_and_late_fail_results_Returns_fail_result()
         {
-            var result = from n in Result<int>.Fail("A")
-                         let n2 = n * 2
-                         from s in Result<string>.Fail("B")
-                         select s;
+            var result =
+                from n in Result<int>.Fail("A")
+                let n2 = n * 2
+                from s in Result<string>.Fail("B")
+                select s;
 
             result.IsFail.Should().BeTrue();
             result.Error.Message.Should().Be("A");
@@ -113,13 +117,13 @@ public class Linq_extension_methods
     public class For_Maybe_of_T
     {
         [Fact]
-        public void Given_all_success_results_and_successful_where_filter_Returns_success_result()
+        public void Given_all_success_results_Returns_success_result()
         {
-            var result = from n in 1.ToMaybe()
-                         let n2 = n * 2
-                         where n2 < 3
-                         from s in $"n2: {n2}".ToMaybe()
-                         select s;
+            var result =
+                from n in 1.ToMaybe()
+                let n2 = n * 2
+                from s in $"n2: {n2}".ToMaybe()
+                select s;
 
             result.IsSuccess.Should().BeTrue();
             result.Value.Should().Be("n2: 2");
@@ -128,11 +132,11 @@ public class Linq_extension_methods
         [Fact]
         public void Given_early_fail_result_Returns_fail_result()
         {
-            var result = from n in Maybe<int>.Fail("A")
-                         let n2 = n * 2
-                         where n2 < 3
-                         from s in $"n2: {n2}".ToMaybe()
-                         select s;
+            var result =
+                from n in Maybe<int>.Fail("A")
+                let n2 = n * 2
+                from s in $"n2: {n2}".ToMaybe()
+                select s;
 
             result.IsFail.Should().BeTrue();
             result.Error.Message.Should().Be("A");
@@ -141,11 +145,11 @@ public class Linq_extension_methods
         [Fact]
         public void Given_early_none_result_Returns_none_result()
         {
-            var result = from n in Maybe<int>.None
-                         let n2 = n * 2
-                         where n2 < 3
-                         from s in $"n2: {n2}".ToMaybe()
-                         select s;
+            var result =
+                from n in Maybe<int>.None
+                let n2 = n * 2
+                from s in $"n2: {n2}".ToMaybe()
+                select s;
 
             result.IsNone.Should().BeTrue();
         }
@@ -153,11 +157,11 @@ public class Linq_extension_methods
         [Fact]
         public void Given_late_fail_result_Returns_fail_result()
         {
-            var result = from n in 1.ToMaybe()
-                         let n2 = n * 2
-                         where n2 < 3
-                         from s in Maybe<string>.Fail("B")
-                         select s;
+            var result =
+                from n in 1.ToMaybe()
+                let n2 = n * 2
+                from s in Maybe<string>.Fail("B")
+                select s;
 
             result.IsFail.Should().BeTrue();
             result.Error.Message.Should().Be("B");
@@ -166,48 +170,36 @@ public class Linq_extension_methods
         [Fact]
         public void Given_late_none_result_Returns_none_result()
         {
-            var result = from n in 1.ToMaybe()
-                         let n2 = n * 2
-                         where n2 < 3
-                         from s in Maybe<string>.None
-                         select s;
+            var result =
+                from n in 1.ToMaybe()
+                let n2 = n * 2
+                from s in Maybe<string>.None
+                select s;
 
             result.IsNone.Should().BeTrue();
         }
 
         [Fact]
-        public void Given_where_filter_returns_false_Returns_none_result()
+        public void Given_early_and_late_fail_results_Returns_fail_result()
         {
-            var result = from n in 1.ToMaybe()
-                         let n2 = n * 2
-                         where n2 < 1
-                         from s in $"n2: {n2}".ToMaybe()
-                         select s;
-
-            result.IsNone.Should().BeTrue();
-        }
-
-        [Fact]
-        public void Given_early_fail_result_and_where_filter_returns_false_Returns_fail_result()
-        {
-            var result = from n in Maybe<int>.Fail("A")
-                         let n2 = n * 2
-                         where n2 < 1
-                         from s in $"n2: {n2}".ToMaybe()
-                         select s;
+            var result =
+                from r1 in Maybe<int>.Fail("A")
+                let abc = "abc"
+                from r2 in Maybe<int>.Fail("B")
+                select abc;
 
             result.IsFail.Should().BeTrue();
             result.Error.Message.Should().Be("A");
         }
 
         [Fact]
-        public void Given_late_fail_result_and_where_filter_returns_false_Returns_none_result()
+        public void Given_early_and_late_none_results_Returns_none_result()
         {
-            var result = from n in 1.ToMaybe()
-                         let n2 = n * 2
-                         where n2 < 1
-                         from s in Maybe<string>.Fail("B")
-                         select s;
+            var result =
+                from r1 in Maybe<int>.None
+                let abc = "abc"
+                from r2 in Maybe<int>.None
+                select abc;
 
             result.IsNone.Should().BeTrue();
         }